sourcesdk: pray that this guess is right

+++ source/sourcesdk/cnetchan.cpp
@@ -2877,10 +2877,10 @@ bool CNetChan::HasPendingReliableData( void )
 		   (m_WaitingList[FRAG_FILE_STREAM].Count() > 0);
 }
 
-float CNetChan::GetTimeConnected() const
+double CNetChan::GetTimeConnected() const
 {
 	double t = net_time - connect_time;
-	return (float)((t>0.0) ? t : 0.0);
+	return (t>0.0) ? t : 0.0;
 }
 
 const netadr_t & CNetChan::GetRemoteAddress() const
@@ -2914,10 +2914,10 @@ float CNetChan::GetTimeoutSeconds() const
 	return m_Timeout;
 }
 
-float CNetChan::GetTimeSinceLastReceived() const
+double CNetChan::GetTimeSinceLastReceived() const
 {
 	double t = net_time - last_received;
-	return (float)((t>0.0) ? t : 0.0);
+	return (t>0.0) ? t : 0.0;
 }
 
 bool CNetChan::IsOverflowed() const
@@ -3042,9 +3042,9 @@ float CNetChan::GetAvgLoss( int flow ) const
 	return m_DataFlow[flow].avgloss;
 }
 
-float CNetChan::GetTime( void ) const
+double CNetChan::GetTime( void ) const
 {
-	return (float)net_time;
+	return net_time;
 }
 
 bool CNetChan::GetStreamProgress( int flow, int *received, int *total ) const
